Non-breaking option to fix preview URLs (#213)
Hostnames are case-insensitive at the DNS level, so the URL class
automatically converts the hostname to lowercase. But as a result,
preview URLs for sandbox IDs with uppercase characters was being routed
to incorrect DO instances.

Add opt-in notmalizeId option that normalizes IDs to lowercase. This is
done in order to prevent existing services from not being able to access
sandbox instances started by an older version. We instead raise warnings
in logs and errors right now so there is no breaking change for anyone,
but giving developers sufficient time to transition to the new behaviour
before we make this the default.